🚚 move: 이미지 업로드 스크린 분리
